+++ CHANGELOG.rst
@@ -1,10 +1,14 @@
 Changelog
 ---------
 
+v0.4.16
+=======
+- Channels (django) (https://github.com/django/channels) server support
+
 v0.4.15
 =======
-- Websockets (https://github.com/python-websockets/websockets) server support
-- Channels (django) (https://github.com/django/channels) server support
+- Websockets server support (https://github.com/python-websockets/websockets)
+- AsyncWebsockets client support (https://github.com/Fuyukai/asyncwebsockets)
 
 v0.4.14
 =======